chore(preload): split preload across files
